refactor(solver): record depth guard verdicts
