+++ crates/tsz-solver/src/evaluation/evaluate_rules/conditional.rs
@@ -3,6 +3,7 @@
 //! Handles TypeScript's conditional types: `T extends U ? X : Y`
 //! Including distributive conditional types over union types.
 
+mod application_infer;
 mod application_reduction;
 mod array_infer;
 mod object_infer;
@@ -321,7 +322,7 @@ impl<'a, R: TypeResolver> TypeEvaluator<'a, R> {
             // Flag TS2589 and return ERROR to prevent stack overflow.
             // This matches tsc's tail recursion limit of 1000 (instantiationCount).
             if tail_recursion_count >= Self::MAX_TAIL_RECURSION_DEPTH {
-                self.mark_depth_exceeded();
+                self.mark_depth_exceeded_for_request();
                 return TypeId::ERROR;
             }
 
@@ -336,7 +337,7 @@ impl<'a, R: TypeResolver> TypeEvaluator<'a, R> {
                     cond.false_type,
                 ))
             {
-                self.mark_depth_exceeded();
+                self.mark_depth_exceeded_for_request();
                 return TypeId::ERROR;
             }
 
@@ -1809,7 +1810,7 @@ impl<'a, R: TypeResolver> TypeEvaluator<'a, R> {
     ) -> TypeId {
         // Limit distribution to prevent OOM with pathologically large unions.
         if members.len() > MAX_CONDITIONAL_DISTRIBUTION_SIZE {
-            self.mark_depth_exceeded();
+            self.mark_depth_exceeded_for_request();
             return TypeId::ERROR;
         }
 
@@ -1903,177 +1904,6 @@ impl<'a, R: TypeResolver> TypeEvaluator<'a, R> {
 
         // Combine results into a union
         self.interner().union_from_slice(&results)
-    }
-
-    /// Try to match conditional types at the Application level before structural expansion.
-    ///
-    /// When both `check_type` and `extends_type` are Applications with the same base type
-    /// (e.g., `Promise<string>` vs `Promise<infer U>`), we can match type arguments
-    /// directly without expanding the interface structure. This is critical for complex
-    /// generic interfaces like Promise, Map, Set where structural expansion makes the
-    /// infer pattern matching fail.
-    fn try_application_infer_match(&mut self, cond: &ConditionalType) -> Option<TypeId> {
-        // Only proceed if extends_type is an Application containing infer.
-        // Keep extends_type as-is (unevaluated) so match_infer_pattern can handle
-        // it at the Application level. This is critical for complex generic interfaces
-        // like Promise, Map, Set where structural expansion loses the ability to
-        // match type arguments directly.
-        let Some(TypeData::Application(pattern_app_id)) = self.interner().lookup(cond.extends_type)
-        else {
-            return None;
-        };
-        let pattern_base = self.interner().type_application(pattern_app_id).base;
-
-        let contains_infer =
-            if let Some(contains_infer) = self.cached_contains_infer(cond.extends_type) {
-                contains_infer
-            } else {
-                let contains_infer = self.type_contains_infer(cond.extends_type);
-                self.cache_contains_infer(cond.extends_type, contains_infer);
-                contains_infer
-            };
-        if !contains_infer {
-            return None;
-        }
-
-        // Recover an Application form for `check_type` whose base matches
-        // `pattern_base`. Three shapes need recovery:
-        //   1. raw type isn't an Application (e.g. `S[K]` inside a per-key
-        //      conditional) — evaluate may yield one;
-        //   2. raw type evaluates to a structural Object/Callable — the
-        //      `display_alias` map records a back-reference to the original
-        //      Application;
-        //   3. raw type IS an Application but its base differs from the
-        //      pattern's (e.g. `Exclude<X<T> | undefined, undefined>` wraps
-        //      `X<T>`) — evaluate through the wrapper so the
-        //      Application-vs-Application match has a same-base source.
-        let mut check_type = cond.check_type;
-        if get_application_base(self.interner(), check_type) != Some(pattern_base) {
-            let evaluated = self.evaluate(check_type);
-            if get_application_base(self.interner(), evaluated) == Some(pattern_base) {
-                check_type = evaluated;
-            } else if let Some(origin) = self.try_recover_application_from_display_alias(evaluated)
-                && get_application_base(self.interner(), origin) == Some(pattern_base)
-            {
-                check_type = origin;
-            }
-        }
-
-        // Skip for special types
-        if check_type == TypeId::ANY || check_type == TypeId::NEVER {
-            return None;
-        }
-        if matches!(
-            self.interner().lookup(check_type),
-            Some(TypeData::TypeParameter(_))
-        ) {
-            return None;
-        }
-
-        // Try infer pattern matching with unevaluated types.
-        // match_infer_pattern handles Application vs Application matching
-        // by comparing base types and recursing on type arguments.
-        let mut checker = self.conditional_subtype_checker();
-        checker.allow_bivariant_rest = true;
-        let mut bindings = FxHashMap::default();
-        let mut visited = InferPatternVisited::default();
-        let matched = self.match_infer_pattern(
-            check_type,
-            cond.extends_type,
-            &mut bindings,
-            &mut visited,
-            &mut checker,
-        );
-        if matched && !bindings.is_empty() {
-            let substituted_true = self.substitute_infer(cond.true_type, &bindings);
-            return Some(self.evaluate(substituted_true));
-        }
-        if self.application_infer_bases_match(check_type, cond.extends_type, &mut checker) {
-            return Some(self.evaluate(cond.false_type));
-        }
-
-        // Last-chance recovery: reduce the source through generic-alias bodies
-        // whose alias body is a conditional that yields an Application form
-        // matching the pattern's base. Handles `Application(ReturnType, [F])
-        // extends Application(Promise, [infer T])` by simulating ReturnType's
-        // body conditional to discover its `Application(Promise, [...])`
-        // substituted true-branch, which the structural fallback cannot
-        // recover from the fully expanded structural object.
-        //
-        // Only worth attempting when the raw source is itself an `Application`
-        // (potentially reducible by alias peeling) or has a display-alias
-        // back-reference to one (recorded for parametric structural bodies).
-        // For intrinsics, type parameters, unions, and other shapes the
-        // reducer would just do one no-op lookup before returning None.
-        for candidate in [cond.check_type, check_type] {
-            if Self::is_alias_reducible_candidate(self.interner(), candidate)
-                && let Some(reduced) = self.reduce_alias_body_to_application_form(candidate)
-                && reduced != candidate
-                && reduced != check_type
-            {
-                let mut checker = self.conditional_subtype_checker();
-                checker.allow_bivariant_rest = true;
-                let mut bindings = FxHashMap::default();
-                let mut visited = InferPatternVisited::default();
-                let matched = self.match_infer_pattern(
-                    reduced,
-                    cond.extends_type,
-                    &mut bindings,
-                    &mut visited,
-                    &mut checker,
-                );
-                if matched && !bindings.is_empty() {
-                    let substituted_true = self.substitute_infer(cond.true_type, &bindings);
-                    return Some(self.evaluate(substituted_true));
-                }
-            }
-        }
-
-        if let Some(alias) = self.try_recover_application_from_display_alias(check_type)
-            && alias != check_type
-        {
-            let mut checker = self.conditional_subtype_checker();
-            checker.allow_bivariant_rest = true;
-            let mut bindings = FxHashMap::default();
-            let mut visited = InferPatternVisited::default();
-            let matched = self.match_infer_pattern(
-                alias,
-                cond.extends_type,
-                &mut bindings,
-                &mut visited,
-                &mut checker,
-            );
-            if matched && !bindings.is_empty() {
-                let substituted_true = self.substitute_infer(cond.true_type, &bindings);
-                return Some(self.evaluate(substituted_true));
-            }
-        }
-
-        None
-    }
-
-    fn application_infer_bases_match(
-        &self,
-        check_type: TypeId,
-        extends_type: TypeId,
-        checker: &mut SubtypeChecker<'_, R>,
-    ) -> bool {
-        let (
-            Some(TypeData::Application(check_app_id)),
-            Some(TypeData::Application(pattern_app_id)),
-        ) = (
-            self.interner().lookup(check_type),
-            self.interner().lookup(extends_type),
-        )
-        else {
-            return false;
-        };
-        let check_app = self.interner().type_application(check_app_id);
-        let pattern_app = self.interner().type_application(pattern_app_id);
-        check_app.args.len() == pattern_app.args.len()
-            && (check_app.base == pattern_app.base
-                || (checker.is_subtype_of(check_app.base, pattern_app.base)
-                    && checker.is_subtype_of(pattern_app.base, check_app.base)))
     }
 }
